refactor: use coursesSlide component instead of popular course and refer course
