--- src/components/PopularCourse/styled.js
@@ -1,5 +1,0 @@
-import styled from 'styled-components'
-
-export const StyledPopularCourse = styled.div`
-
-`
