--- src/components/ReferCourse/styled.js
@@ -1,5 +1,0 @@
-import styled from 'styled-components'
-
-export const StyledReferCourse = styled.div`
-  
-`
