+++ src/components/SlideCourses/styled.js
