+++ src/components/index.js
@@ -2,14 +2,12 @@ import Header from './Header'
 import Footer from './Footer'
 import Toast from './Toast'
 import CardGlobal from './CardGlobal'
-import SwiperCarosel from './Swiper'
-import ReferCourse from './ReferCourse'
-import PopularCourse from './PopularCourse'
 import ErrorBoundary from './ErrorBoundary'
 import Loading from './Loading'
+import SlideCourses from './SlideCourses'
 
 export * from './FormControl'
 export {
-  Header, Footer, Toast, CardGlobal, SwiperCarosel, ReferCourse, PopularCourse,
-  ErrorBoundary, Loading
+  Header, Footer, Toast, CardGlobal,
+  ErrorBoundary, Loading, SlideCourses
 }
